apps/backend/core/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.7
django-cors-headers==4.3.1
django-crontab==0.7.1
django-dbbackup==4.0.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
kombu==5.3.5
mysqlclient==2.2.1
packaging==23.2
prompt-toolkit==3.0.43
python-benedict==0.33.1
python-dateutil==2.8.2
python-fsutil==0.13.0
python-slugify==8.0.1
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
